fix: sample app
fix: fatal hang on ios
