+++ examples/default/android/build.gradle
@@ -24,7 +24,7 @@ buildscript {
         classpath("com.android.tools.build:gradle:8.1.0")
         classpath("com.facebook.react:react-native-gradle-plugin")
         classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion")
-        classpath("com.instabug.library:instabug-plugin:15.0.1")
+        classpath("com.instabug.library:instabug-plugin:16.0.0")
     }
 }
