+++ plugin/src/pluginProps.ts
@@ -1,0 +1,6 @@
+export interface PluginProps {
+  name?: string;
+  forceUploadSourceMaps?: boolean;
+  addScreenRecordingBugReportingPermission?: boolean;
+  addBugReportingIosMediaPermission?: boolean;
+}
